pyproject.toml
[project]
name = "smm-panel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-multipart",
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
